modularize DDPG network
